Correções no crud e adição de tratativas extras em praticamente todos os arquivos

+++ Materia/proc_del_materia.php
@@ -2,14 +2,28 @@
     require_once("../Componente/header.php");
     require_once("materia.php");
     require_once("../Utils/valida_formulario.php");
-    require_once("../Materia/materia.php"); //verificação da exclusão
+    require_once("../Matricula/matricula.php");
+    require_once("../Avaliacao/avaliacao.php");
 
     if (isset($_POST['idmateriaDEL'])){
-        $id = trim($_POST['idmateriaDEL']);
-        // liberarExclusao($id)          
-                deleteMateria($id);
-                header("Location: form_materia.php?del=1");
-                }
+        $id = $_POST['idmateriaDEL'];
+
+        if (!is_numeric($id)){
+            header("Location: form_materia.php?del=2");
+            exit;
+        }
+
+        if (existeMatriculaMateria($id) || existeAvaliacaoMateria($id)){
+            header("Location: form_materia.php?del=0");
+            exit;
+        }
+
+        deleteMateria($id);
+        header("Location: form_materia.php?del=1");
+        exit;
+    }
+
+    header("Location: form_materia.php");
 ?>
